+++ RELEASES.md
@@ -1,5 +1,13 @@
 # Release notes
 
+## unreleased (2023-XX-XX)
+
+### Added
+
+- `views::SubGraph` for views into non-hierarchical subgraphs ([#100][])
+
+  [#100]: https://github.com/CQCL/portgraph/issues/100
+
 ## v0.7.2 (2023-07-31)
 
 ### Added